Move src/scripts/runtime to src/runtime
