+++ src/scripts/mod.rs
@@ -2,4 +2,3 @@ pub mod ctx;
 pub use self::ctx::Script;
 pub mod loader;
 pub use self::loader::Loader;
-pub mod runtime;
